% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarity_api_r6.R
\docType{data}
\name{dirs}
\alias{dirs}
\title{Default directory tree for specific data types}
\format{
An object of class \code{list} of length 5.
}
\usage{
dirs
}
\description{
COHHIO specific directory tree \emph{Recommended}
}
\keyword{datasets}
