% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_export.R
\docType{data}
\name{.hud_export}
\alias{.hud_export}
\title{HUD Export saved values}
\format{
An object of class \code{list} of length 23.
}
\usage{
.hud_export
}
\description{
The Look Ids for various filter criteria and the column specifications
}
\keyword{datasets}
