% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_export_r6.R
\name{fetch}
\alias{fetch}
\title{Retrieve data from disk or the API}
\usage{
fetch(x, look_type, path, .write = FALSE, ee)
}
\arguments{
\item{x}{\code{(character)} The HUD CSV Export item name}

\item{look_type}{\code{(character)} The look type to retrieve. One of:
\itemize{
\item{year2}{Two Complete Years}
\item{s2020}{Since the beginning of 2020}
\item{daily}{Created or updated in the last complete day}
}}

\item{path}{\code{(character)} The directory path in which to search}

\item{write}{\code{(logical)} Whether to write the raw data from the API and the renamed data to the data/API folder}
}
\value{
\code{(tibble)} The HUD Export item requested.
}
\description{
Determines the appropriate location from which to retrieve HUD Export data
}
\section{Export_Items}{

Available HUD Export Items are
\itemize{
\item{Affiliation}
\item{Assessment}
\item{AssessmentQuestions}
\item{AssessmentResults}
\item{Client}
\item{CurrentLivingSituation}
\item{Disabilities}
\item{EmploymentEducation}
\item{Enrollment}
\item{EnrollmentCoC}
\item{Event}
\item{Exit}
\item{Export}
\item{Funder}
\item{HealthAndDV}
\item{IncomeBenefits}
\item{Inventory}
\item{Organization}
\item{Project}
\item{ProjectCoC}
\item{Services}
\item{User}
\item{YouthEducationStatus}
}
}

