% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_clients}
\alias{find_clients}
\title{Filter for specific clients}
\usage{
find_clients(x, clients)
}
\arguments{
\item{x}{\code{(data.frame)} with either UniqueID or PersonalID}

\item{clients}{\code{(character)} of either UniqueID's or PersonalID's. The corresponding column must be present in the data to filter}
}
\value{
\code{(data.frame)} filtered for clients
}
\description{
Filter for specific clients
}
