% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_projects}
\alias{find_projects}
\title{Filter for specific projects}
\usage{
find_projects(x, projects)
}
\arguments{
\item{x}{\code{(data.frame)} with either ProjectID or ProjectName}

\item{clients}{\code{(character)} of either ProjectID's or ProjectName's. The corresponding column must be present in the data to filter}
}
\value{
\code{(data.frame)} filtered for clients
}
\description{
Filter for specific projects
}
