% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{folder_looks}
\alias{folder_looks}
\title{Retrieve all Look IDs from a folder}
\usage{
folder_looks(folder)
}
\arguments{
\item{folder}{\code{(folder)} folder object}
}
\value{
\code{(named character)} vector of Look IDs
}
\description{
Retrieve the Look IDS from a folder with their corresponding names
}
