% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_export_r6.R
\name{hud_export}
\alias{hud_export}
\title{Call HUD Export Items from the Clarity Looker API}
\description{
Calls the Clarity Looker HUD CSV Export  (BETA) API to return to the HUD Export Items on various time ranges via pre-constructed Looks. See \code{?fetch} for details on using all data methods.
}
\section{Export_Items}{

Available HUD Export Items are
\itemize{
\item{Affiliation}
\item{Assessment}
\item{AssessmentQuestions}
\item{AssessmentResults}
\item{Client}
\item{CurrentLivingSituation}
\item{Disabilities}
\item{EmploymentEducation}
\item{Enrollment}
\item{EnrollmentCoC}
\item{Event}
\item{Exit}
\item{Export}
\item{Funder}
\item{HealthAndDV}
\item{IncomeBenefits}
\item{Inventory}
\item{Organization}
\item{Project}
\item{ProjectCoC}
\item{Services}
\item{User}
\item{YouthEducationStatus}
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_export}{\code{hud_export$get_export()}}
\item \href{#method-update_export}{\code{hud_export$update_export()}}
\item \href{#method-get_extras}{\code{hud_export$get_extras()}}
\item \href{#method-update_extras}{\code{hud_export$update_extras()}}
\item \href{#method-new}{\code{hud_export$new()}}
\item \href{#method-close}{\code{hud_export$close()}}
\item \href{#method-clone}{\code{hud_export$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_export"></a>}}
\if{latex}{\out{\hypertarget{method-get_export}{}}}
\subsection{Method \code{get_export()}}{
Pull all Export items with associate Looks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$get_export(path = self$dirs$export)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_export"></a>}}
\if{latex}{\out{\hypertarget{method-update_export}{}}}
\subsection{Method \code{update_export()}}{
Run daily update for all HUD Export items on disk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$update_export(
  path = self$dirs$export,
  skip = c("Assessment", "AssessmentQuestions", "AssessmentResults",
    "YouthEducationStatus")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_extras"></a>}}
\if{latex}{\out{\hypertarget{method-get_extras}{}}}
\subsection{Method \code{get_extras()}}{
Pull all Extra items with associate Looks
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$get_extras(path = self$dirs$extras)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_extras"></a>}}
\if{latex}{\out{\hypertarget{method-update_extras}{}}}
\subsection{Method \code{update_extras()}}{
Run daily update for all HUD Export items on disk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$update_extras(path = self$dirs$extras, skip = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize the Looker API connection given the path to the ini configuration file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$new(
  configFile,
  dirs = list(export = "data/export", public = "data/public", spm = "data/spm", extras
    = "data/extras")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{configFile}}{\code{(character)} Path to the Looker \emph{.ini} configuration file. Only the directory path is needed if the file is entitled \emph{Looker.ini}}

\item{\code{dirs}}{\code{(named list)} of default directory paths for where to store the feather files for the following data types:
\itemize{
\item{\code{export}}{ The HUD Export items \strong{Default}: \emph{data/export}}
\item{\code{public}}{ The public items \strong{Default}: \emph{data/public}}
\item{\code{spm}}{ The SPM items \strong{Default}: \emph{data/spm}}
\item{\code{extras}}{ The HUD Extras (custom items) \strong{Default}: \emph{data/extras}}
}
This is optional and the path can be provided to individual methods as needed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
Close the Looker API Connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
