% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hud_export_extract}
\alias{hud_export_extract}
\title{Extract a previously downloaded HUD Export archive}
\usage{
hud_export_extract(
  browser_dl_folder = "~/../Downloads",
  extract_path = file.path("data", "export"),
  delete_archive = TRUE,
  moment = Sys.Date(),
  wait = lubridate::minutes(2)
)
}
\arguments{
\item{browser_dl_folder}{\code{(character)} path to the browser's download folder or the file to extract}

\item{extract_path}{\code{(character)} path to the folder where the archive is to be extracted}

\item{delete_archive}{\code{(logical)} Delete the archive after extracting?}

\item{moment}{\code{(POSIXct/Date)} The time point which the archive creation time should be greater than to ensure it's recent.}

\item{wait}{\code{(Duration)} to wait for the file to appear in the download directory. Relevant when using browser automation.}
}
\value{
\code{(logical)} as to whether the extraction was successful
}
\description{
Extract a previously downloaded HUD Export archive
}
