% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_looker.R
\name{hud_export_look_update}
\alias{hud_export_look_update}
\title{Update the \code{.hud_export} spec given a \code{look_table}}
\usage{
hud_export_look_update(look_table, look_type)
}
\arguments{
\item{look_table}{\code{(numeric)} output from \code{parse_look_table}}

\item{look_type}{\code{(character)} which Look type to update (or add)}
}
\value{
Console output from dput with the new spec to copy/paste into hud_export.R
}
\description{
Using the \code{look_table} returned from \code{parse_look_table}, update the \code{.hud_export} spec
}
