% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hud_feather}
\alias{hud_feather}
\title{Write object to the \emph{data} directory as feather file}
\usage{
hud_feather(.data, path = "data", nm)
}
\arguments{
\item{.data}{\code{(tibble/data.frame)} The object to write to feather}

\item{path}{\code{(character)} A single directory path or full file path where the feather file will be saved. If only a directory is provided, the file will be named after the \code{.data} object (e.g., "Affiliation.feather"). If a filename is provided in the path (e.g., "Export.feather"), the file will be saved with that name.}

\item{nm}{\code{(character)} Optional name for the file. If not provided, the name of the \code{.data} object will be used.}
}
\value{
A success message at the console
}
\description{
Writes a \code{tibble/data.frame} as a feather file to the path
}
