% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hud_load}
\alias{hud_load}
\title{Load a HUD Export item from disk}
\usage{
hud_load(x, path = "data")
}
\arguments{
\item{x}{\code{(character)} The HUD CSV Export item name}

\item{path}{\code{(character)} The directory path in which to search}
}
\value{
\code{(POSIXct)} Last modified time
}
\description{
Load the named HUD Export item from the \code{path} provided
}
\section{Export_Items}{

Available HUD Export Items are
\itemize{
\item{Affiliation}
\item{Assessment}
\item{AssessmentQuestions}
\item{AssessmentResults}
\item{Client}
\item{CurrentLivingSituation}
\item{Disabilities}
\item{EmploymentEducation}
\item{Enrollment}
\item{EnrollmentCoC}
\item{Event}
\item{Exit}
\item{Export}
\item{Funder}
\item{HealthAndDV}
\item{IncomeBenefits}
\item{Inventory}
\item{Organization}
\item{Project}
\item{ProjectCoC}
\item{Services}
\item{User}
\item{YouthEducationStatus}
}
}

