% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_c.R
\name{knn_dist.default}
\alias{knn_dist.default}
\title{Default Method for K-Nearest Neighbors Distance}
\usage{
\method{knn_dist}{default}(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A numeric vector of distances between the elements of \code{x} and \code{y}, calculated as \code{log(x - y)}.
}
\description{
Computes the distance between two numeric vectors using the logarithmic difference.
}
\details{
This method calculates the logarithmic difference between elements of \code{x} and \code{y}. Note that this method assumes \code{x} and \code{y} are numeric and may produce warnings if there are negative values or zero.
}
