% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_c.R
\name{knn_dist.factor}
\alias{knn_dist.factor}
\title{K-Nearest Neighbors Distance for Factor Vectors}
\usage{
\method{knn_dist}{factor}(x, y)
}
\arguments{
\item{x}{A factor vector.}

\item{y}{A factor vector.}
}
\value{
A numeric vector of distances between the elements of \code{x} and \code{y}, calculated using \code{stringdist::stringdist} on their character representations.
}
\description{
Computes the distance between two factor vectors by converting them to character vectors.
}
\details{
This method first converts the factors to character vectors and then computes distances using \code{stringdist::stringdist}.
}
