% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_c.R
\name{knn_dist}
\alias{knn_dist}
\title{K-Nearest Neighbors Distance Generic Function}
\usage{
knn_dist(x, y)
}
\arguments{
\item{x}{The first argument for method dispatch. The type of this argument determines which method is called.}

\item{y}{The second argument for method dispatch. The type of this argument determines which method is called.}
}
\value{
The return value depends on the method implemented for the specific classes of \code{x} and \code{y}.
}
\description{
A generic function for computing distances between data points for k-nearest neighbors analysis.
}
\details{
This function is a generic for computing k-nearest neighbors distances. Methods should be implemented for specific classes to define how distances are calculated.
}
\seealso{
\code{\link{knn_dist.default}} for the default method.
}
