% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{look_id_from_folder}
\alias{look_id_from_folder}
\title{Retrieve Look info from a folder}
\usage{
look_id_from_folder(looks, folder)
}
\arguments{
\item{looks}{\code{(character)} of the Look names for which to retrieve info for}

\item{folder}{\code{(folder)} folder object}
}
\value{
\code{(named list)} of Look Infos
}
\description{
Retrieve Look Info from a folder for a specific look name
}
