% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_link}
\alias{make_link}
\title{Make a Clarity link using an \code{ID} column}
\usage{
make_link(ID, link_text, type = NULL, chr = TRUE)
}
\arguments{
\item{ID}{\code{(character)} The \code{ID} column. See the list below for the appropriate matching \code{ID}. If \code{link_text} is:
\itemize{
\item{\code{UniqueID}}{ \code{PersonalID}}
\item{\code{EnrollmentID}}{ \code{PersonalID}}
\item{\code{ProjectName}}{ \code{ProjectID}}
\item{\code{AgencyName}}{ \code{AgencyID}}
}}

\item{link_text}{\code{(character)} The link text. Typically \code{UniqueID/EnrollmentID/AgencyName/ProjectName} column}

\item{type}{\code{(character)} the type of link to create based on \code{link_text}. One of:
\itemize{
\item{\code{"profile"}}{ When \code{link_text} is \code{UniqueID}}
\item{\code{"enrollment"}}{ When \code{link_text} is \code{EnrollmentID}}
\item{\code{"agency_switch"}}{ When \code{link_text} is \code{AgencyName}}
\item{\code{"program_edit"}}{ When \code{link_text} is \code{ProgramName}}
}}

\item{chr}{\code{(logical)} Whether to output a character or a \code{shiny.tag} if \code{FALSE}. \strong{Default} TRUE}
}
\value{
\code{(character/shiny.tag)} If \code{chr = TRUE} a character vector, if \code{chr = FALSE} a \code{shiny.tag}.
}
\description{
If used in a \link[DT]{datatable}, set \code{escape = FALSE}
}
\examples{
data.frame(a = letters, b = seq_along(letters)) |>  dplyr::mutate(a = make_link(ID = b, link_text = a, type = "profile"))
}
