% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_looker.R
\name{parse_look_table}
\alias{parse_look_table}
\title{Extract Looks and their names from the folder browse table in Looker}
\usage{
parse_look_table(htm)
}
\arguments{
\item{htm}{\code{(character)} Path to html file with outerHTML copy/pasted into it}
}
\value{
\code{(numeric)} vector of look ids and their respective names
}
\description{
Takes file with the "outerHTML" from chrome > inspect of the table of all the looks in Looker folder browse view and parse it to extract the look id numbers and their names provided that the format of the look name is HUD_Export.\link{ExportItemName}
}
