% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_looker.R
\name{setup_looker}
\alias{setup_looker}
\title{Setup the Looker ini file}
\usage{
setup_looker(
  client_id,
  client_secret,
  ini_filepath = "Looker.ini",
  base_url = "https://looker.clarityhs.com:19999"
)
}
\arguments{
\item{client_id}{\code{(character)} The Client ID provided by Clarity}

\item{client_secret}{\code{(character)} The Client Secret provided by Clarity}

\item{ini_filepath}{\code{(character)} The directory and name of the ini file. \strong{Default} creates \emph{Looker.ini} in the working directory. A path can also be provided, and the file will be named \emph{Looker.ini} in that directory.}

\item{base_url}{\code{(character)} The base URL to the Clarity Looker instance. This is typically the same for all Clarity users.}
}
\description{
Connecting to the Looker API via \href{https://github.com/looker-open-source/lookr}{\code{lookr}} requires an ini file with credentials to access the API. This function creates that ini file in a directory of your choosing.
}
