% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_export.R
\name{to_feather}
\alias{to_feather}
\title{Write object to the \emph{data} directory}
\usage{
to_feather(x, path = "data")
}
\arguments{
\item{x}{\code{(tibble/data.frame)} The object to write to feather}

\item{path}{\code{(character vector)} A character vector of the directory path to be passed to \link[base]{file.path}}
}
\value{
A success message at the console
}
\description{
Writes a \code{tibble/data.frame} as a feather file to the \emph{data} directory using the name of the object as the file name.
}
