% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Client_filter}
\alias{Client_filter}
\title{Filter out specific Clients}
\usage{
Client_filter(x, clients = getOption("HMIS")$clients_to_filter)
}
\arguments{
\item{x}{\code{(data.frame)} With PersonalID or UniqueID column}

\item{clients}{\code{(character)} of PersonalIDs to filter with names corresponding their UniqueIDs (Clarity only)}
}
\value{
\code{(data.frame)} without \code{clients_to_filter}
}
\description{
Often used to filter test/training demo clients
}
\concept{Client functions}
