% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_export_r6.R
\name{hud_export}
\alias{hud_export}
\title{Call HUD Export Items from the Clarity Looker API}
\description{
Calls the Clarity Looker HUD CSV Export  (BETA) API to return to the HUD Export Items on various time ranges via pre-constructed Looks.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Affiliation}{\code{hud_export$Affiliation()}}
\item \href{#method-Assessment}{\code{hud_export$Assessment()}}
\item \href{#method-AssessmentQuestions}{\code{hud_export$AssessmentQuestions()}}
\item \href{#method-AssessmentResults}{\code{hud_export$AssessmentResults()}}
\item \href{#method-Client}{\code{hud_export$Client()}}
\item \href{#method-CurrentLivingSituation}{\code{hud_export$CurrentLivingSituation()}}
\item \href{#method-Disabilities}{\code{hud_export$Disabilities()}}
\item \href{#method-EmploymentEducation}{\code{hud_export$EmploymentEducation()}}
\item \href{#method-Enrollment}{\code{hud_export$Enrollment()}}
\item \href{#method-EnrollmentCoC}{\code{hud_export$EnrollmentCoC()}}
\item \href{#method-Event}{\code{hud_export$Event()}}
\item \href{#method-Exit}{\code{hud_export$Exit()}}
\item \href{#method-Export}{\code{hud_export$Export()}}
\item \href{#method-Funder}{\code{hud_export$Funder()}}
\item \href{#method-HealthAndDV}{\code{hud_export$HealthAndDV()}}
\item \href{#method-IncomeBenefits}{\code{hud_export$IncomeBenefits()}}
\item \href{#method-Inventory}{\code{hud_export$Inventory()}}
\item \href{#method-Organization}{\code{hud_export$Organization()}}
\item \href{#method-Project}{\code{hud_export$Project()}}
\item \href{#method-ProjectCoC}{\code{hud_export$ProjectCoC()}}
\item \href{#method-Services}{\code{hud_export$Services()}}
\item \href{#method-User}{\code{hud_export$User()}}
\item \href{#method-YouthEducationStatus}{\code{hud_export$YouthEducationStatus()}}
\item \href{#method-new}{\code{hud_export$new()}}
\item \href{#method-close}{\code{hud_export$close()}}
\item \href{#method-clone}{\code{hud_export$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Affiliation"></a>}}
\if{latex}{\out{\hypertarget{method-Affiliation}{}}}
\subsection{Method \code{Affiliation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Affiliation(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Assessment"></a>}}
\if{latex}{\out{\hypertarget{method-Assessment}{}}}
\subsection{Method \code{Assessment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Assessment(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssessmentQuestions"></a>}}
\if{latex}{\out{\hypertarget{method-AssessmentQuestions}{}}}
\subsection{Method \code{AssessmentQuestions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$AssessmentQuestions(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AssessmentResults"></a>}}
\if{latex}{\out{\hypertarget{method-AssessmentResults}{}}}
\subsection{Method \code{AssessmentResults()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$AssessmentResults(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Client"></a>}}
\if{latex}{\out{\hypertarget{method-Client}{}}}
\subsection{Method \code{Client()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Client(look_type = "disk", path = file.path("data"), write = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CurrentLivingSituation"></a>}}
\if{latex}{\out{\hypertarget{method-CurrentLivingSituation}{}}}
\subsection{Method \code{CurrentLivingSituation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$CurrentLivingSituation(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Disabilities"></a>}}
\if{latex}{\out{\hypertarget{method-Disabilities}{}}}
\subsection{Method \code{Disabilities()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Disabilities(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmploymentEducation"></a>}}
\if{latex}{\out{\hypertarget{method-EmploymentEducation}{}}}
\subsection{Method \code{EmploymentEducation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$EmploymentEducation(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Enrollment"></a>}}
\if{latex}{\out{\hypertarget{method-Enrollment}{}}}
\subsection{Method \code{Enrollment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Enrollment(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EnrollmentCoC"></a>}}
\if{latex}{\out{\hypertarget{method-EnrollmentCoC}{}}}
\subsection{Method \code{EnrollmentCoC()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$EnrollmentCoC(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Event"></a>}}
\if{latex}{\out{\hypertarget{method-Event}{}}}
\subsection{Method \code{Event()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Event(look_type = "disk", path = file.path("data"), write = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exit"></a>}}
\if{latex}{\out{\hypertarget{method-Exit}{}}}
\subsection{Method \code{Exit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Exit(look_type = "disk", path = file.path("data"), write = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Export"></a>}}
\if{latex}{\out{\hypertarget{method-Export}{}}}
\subsection{Method \code{Export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Export(look_type = "disk", path = file.path("data"), write = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Funder"></a>}}
\if{latex}{\out{\hypertarget{method-Funder}{}}}
\subsection{Method \code{Funder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Funder(look_type = "disk", path = file.path("data"), write = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HealthAndDV"></a>}}
\if{latex}{\out{\hypertarget{method-HealthAndDV}{}}}
\subsection{Method \code{HealthAndDV()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$HealthAndDV(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncomeBenefits"></a>}}
\if{latex}{\out{\hypertarget{method-IncomeBenefits}{}}}
\subsection{Method \code{IncomeBenefits()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$IncomeBenefits(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Inventory"></a>}}
\if{latex}{\out{\hypertarget{method-Inventory}{}}}
\subsection{Method \code{Inventory()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Inventory(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Organization"></a>}}
\if{latex}{\out{\hypertarget{method-Organization}{}}}
\subsection{Method \code{Organization()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Organization(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project"></a>}}
\if{latex}{\out{\hypertarget{method-Project}{}}}
\subsection{Method \code{Project()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Project(look_type = "disk", path = file.path("data"), write = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProjectCoC"></a>}}
\if{latex}{\out{\hypertarget{method-ProjectCoC}{}}}
\subsection{Method \code{ProjectCoC()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$ProjectCoC(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Services"></a>}}
\if{latex}{\out{\hypertarget{method-Services}{}}}
\subsection{Method \code{Services()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$Services(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-User"></a>}}
\if{latex}{\out{\hypertarget{method-User}{}}}
\subsection{Method \code{User()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$User(look_type = "disk", path = file.path("data"), write = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-YouthEducationStatus"></a>}}
\if{latex}{\out{\hypertarget{method-YouthEducationStatus}{}}}
\subsection{Method \code{YouthEducationStatus()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$YouthEducationStatus(
  look_type = "disk",
  path = file.path("data"),
  write = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize the Looker API connection given the path to the ini configuration file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$new(configFile)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{configFile}}{\code{(character)} Path to the Looker \emph{.ini} configuration file. Only the directory path is needed if the file is entitled \emph{Looker.ini}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
Close the Looker API Connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hud_export$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
