% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hud_feather}
\alias{hud_feather}
\title{Write object to the \emph{data} directory as feather file}
\usage{
hud_feather(.data, path = "data")
}
\arguments{
\item{.data}{\code{(tibble/data.frame)} The object to write to feather}

\item{path}{\code{(character vector)} A character vector of the directory path to be passed to \link[base]{file.path}. If the filename is appended IE "Export.feather", the file will be named as such. Otherwise the name of the \code{.data} object will be used IE, if an object named \code{Affiliation} containing the Affiliation data is passed to \code{.data}, the resulting file will be \link{path}/Affiliation.feather.}
}
\value{
A success message at the console
}
\description{
Writes a \code{tibble/data.frame} as a feather file to the path
}
