% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_link}
\alias{make_link}
\title{Make a Clarity link using the \code{PersonalID} and \code{UniqueID/EnrollmentID}}
\usage{
make_link(PersonalID, ID, chr = TRUE)
}
\arguments{
\item{PersonalID}{\code{(character)} The \code{PersonalID} column}

\item{ID}{\code{(character)} The \code{UniqueID/EnrollmentID} column}

\item{chr}{\code{(logical)} Whether to output a character or a \code{shiny.tag} if \code{FALSE}. \strong{Default} TRUE}
}
\value{
\code{(character/shiny.tag)} If \code{PersonalID} is a character vector (IE nested in a mutate), and \code{chr = TRUE} a character vector, if \code{chr = FALSE} a \code{shiny.tag}. The \code{ID} column will be replaced with the link.
}
\description{
If used in a \link[DT]{datatable}, set \code{escape = FALSE}
}
\examples{
data.frame(a = letters, b = seq_along(letters)) |>  dplyr::mutate(a = make_link(a, b))
}
